feat(settings): Company section with Profile + Branding tabs (#233)
Slice 7 of the Settings redesign (PRD #226). The Company section gets
two tabs at /settings/company:

- Profile: the existing /settings/company body, extracted into
  profile-tab.tsx unchanged.
- Branding: NEW shape that merges the old /settings/appearance and
  /settings/pdf-presets list pages into a single scrollable layout with
  two sub-sections (NOT nested tabs) — Appearance first, then PDF
  Presets, each under its own h2 heading. Rationale (PRD-locked): a
  normal user thinks "how my business looks" without distinguishing
  app-theme from document-look; merging avoids bouncing between two
  settings pages for related visual choices.

Behavior wired through SettingsTabs from Slice 1 (#227). Profile is the
default tab; ?tab=branding selects the second.

Other changes:

- Server-side redirects from /settings/appearance and /settings/pdf-presets
  to /settings/company?tab=branding so bookmarks keep working.
- /settings/pdf-presets/[id]/edit is unchanged; the Branding tab links
  to it as before.
- settings-nav.ts: one "Company" entry (icon Building2) replacing the
  legacy "Company Profile"; the two old Appearance + PDF Presets
  entries are dropped. The layout iterates this list for both the
  desktop sidebar and the mobile dropdown, so both surfaces update in
  one place.
- Smoke test asserts /settings/company renders Profile + Branding tab
  labels in order; settings-nav.test.ts guards the new nav shape.
