+++ src/lib/settings-nav.test.ts
@@ -45,3 +45,33 @@ describe("settingsNavItems — Money section (slice 4)", () => {
     expect(stale).toEqual([]);
   });
 });
+
+// #233 — guard the canonical settings nav shape after Slice 7. A single
+// "Company" entry replaces "Company Profile" and now owns Appearance and
+// PDF Presets as tabs inside /settings/company. The old top-level entries
+// for those two pages are gone.
+describe("settingsNavItems (Slice 7 — Company + Branding merge)", () => {
+  it("has a single Company entry pointing at /settings/company", () => {
+    const companyEntries = settingsNavItems.filter(
+      (item) => item.href === "/settings/company",
+    );
+    expect(companyEntries).toHaveLength(1);
+    expect(companyEntries[0].label).toBe("Company");
+  });
+
+  it("no longer has top-level Appearance or PDF Presets entries", () => {
+    const appearance = settingsNavItems.find(
+      (item) => item.href === "/settings/appearance",
+    );
+    const pdfPresets = settingsNavItems.find(
+      (item) => item.href === "/settings/pdf-presets",
+    );
+    expect(appearance).toBeUndefined();
+    expect(pdfPresets).toBeUndefined();
+  });
+
+  it("no longer has the legacy 'Company Profile' label", () => {
+    const labels = settingsNavItems.map((item) => item.label);
+    expect(labels).not.toContain("Company Profile");
+  });
+});
